requirements.txt
Flask
python-dotenv
crowdstrike-falconpy
python-dateutil
python-dotenv
pdfkit
weasyprint